refactor(post): move PostSummary from search.dto to post.dto
Colocates the minimal post summary schema with the other PublicPost/
OwnerPost types. search.dto imports it from its canonical home. No
behavior change.
